add a launch queue, split reader funcs
