+++ pkg/daemon/daemon.go
@@ -22,8 +22,6 @@ along with TapTo.  If not, see <http://www.gnu.org/licenses/>.
 package daemon
 
 import (
-	"encoding/hex"
-	"errors"
 	"fmt"
 	"os"
 	"strings"
@@ -31,152 +29,32 @@ import (
 
 	"golang.org/x/exp/slices"
 
-	"github.com/clausecker/nfc/v2"
 	"github.com/rs/zerolog/log"
 	"github.com/wizzomafizzo/mrext/pkg/input"
 	"github.com/wizzomafizzo/tapto/pkg/config"
 	"github.com/wizzomafizzo/tapto/pkg/launcher"
 	"github.com/wizzomafizzo/tapto/pkg/platforms/mister"
-	"github.com/wizzomafizzo/tapto/pkg/tokens"
-	"github.com/wizzomafizzo/tapto/pkg/utils"
 )
 
-const (
-	timeToForgetCard   = 500 * time.Millisecond
-	connectMaxTries    = 10
-	TimesToPoll        = 20
-	PeriodBetweenPolls = 300 * time.Millisecond
-	periodBetweenLoop  = 300 * time.Millisecond
-)
-
-func pollDevice(
-	cfg *config.UserConfig,
-	pnd *nfc.Device,
-	activeCard Token,
-	ttp int,
-	pbp time.Duration,
-) (Token, bool, error) {
-	removed := false
-
-	count, target, err := pnd.InitiatorPollTarget(tokens.SupportedCardTypes, ttp, pbp)
-	if err != nil && !errors.Is(err, nfc.Error(nfc.ETIMEOUT)) {
-		return activeCard, removed, err
-	}
-
-	if count <= 0 {
-		if activeCard.UID != "" && time.Since(activeCard.ScanTime) > timeToForgetCard {
-			log.Info().Msg("card removed")
-			activeCard = Token{}
-			removed = true
-		}
-
-		return activeCard, removed, nil
-	}
-
-	cardUid := tokens.GetCardUID(target)
-	if cardUid == "" {
-		log.Warn().Msgf("unable to detect token UID: %s", target.String())
-	}
-
-	if cardUid == activeCard.UID {
-		return activeCard, removed, nil
-	}
-
-	log.Info().Msgf("found token UID: %s", cardUid)
-
-	var record []byte
-	cardType := tokens.GetCardType(target)
-
-	if cardType == tokens.TypeNTAG {
-		log.Info().Msg("NTAG detected")
-		record, err = tokens.ReadNtag(*pnd)
-		if err != nil {
-			return activeCard, removed, fmt.Errorf("error reading ntag: %s", err)
-		}
-		cardType = tokens.TypeNTAG
-	}
-
-	if cardType == tokens.TypeMifare {
-		log.Info().Msg("Mifare detected")
-		record, err = tokens.ReadMifare(*pnd, cardUid)
-		if err != nil {
-			log.Error().Msgf("error reading mifare: %s", err)
-		}
-		cardType = tokens.TypeMifare
-	}
-
-	log.Debug().Msgf("record bytes: %s", hex.EncodeToString(record))
-	tagText := tokens.ParseRecordText(record)
-	if tagText == "" {
-		log.Warn().Msg("no text NDEF found")
-	} else {
-		log.Info().Msgf("decoded text NDEF: %s", tagText)
-	}
-
-	card := Token{
-		Type:     cardType,
-		UID:      cardUid,
-		Text:     tagText,
-		ScanTime: time.Now(),
-	}
-
-	return card, removed, nil
+type TokenQueue struct {
+	tokens chan Token
 }
 
-func detectConnectionString() string {
-	log.Info().Msg("probing for serial devices")
-	devices, _ := utils.GetLinuxSerialDeviceList()
-
-	for _, device := range devices {
-		connectionString := "pn532_uart:" + device
-		pnd, err := nfc.Open(connectionString)
-		log.Info().Msgf("trying %s", connectionString)
-		if err == nil {
-			log.Info().Msgf("success using serial: %s", connectionString)
-			pnd.Close()
-			return connectionString
-		}
-	}
-
-	return ""
+func (q *TokenQueue) Enqueue(t Token) {
+	q.tokens <- t
 }
 
-func OpenDeviceWithRetries(config config.TapToConfig, state *State) (nfc.Device, error) {
-	var connectionString = config.ConnectionString
-	if connectionString == "" && config.ProbeDevice == true {
-		connectionString = detectConnectionString()
-	}
+func (q *TokenQueue) Dequeue() Token {
+	return <-q.tokens
+}
 
-	log.Info().Msgf("connecting to device: %s", connectionString)
+func (q *TokenQueue) Close() {
+	close(q.tokens)
+}
 
-	tries := 0
-	for {
-		pnd, err := nfc.Open(connectionString)
-		if err == nil {
-			log.Info().Msgf("successful connect after %d tries", tries)
-
-			connProto := strings.SplitN(strings.ToLower(connectionString), ":", 2)[0]
-			log.Info().Msgf("connection protocol: %s", connProto)
-			deviceName := pnd.String()
-			log.Info().Msgf("device name: %s", deviceName)
-
-			if connProto == "pn532_uart" {
-				state.SetReaderConnected(ReaderTypePN532)
-			} else if strings.Contains(deviceName, "ACR122U") {
-				state.SetReaderConnected(ReaderTypeACR122U)
-			} else {
-				state.SetReaderConnected(ReaderTypeUnknown)
-			}
-
-			return pnd, err
-		}
-
-		if tries >= connectMaxTries {
-			log.Error().Msgf("could not open device after %d tries: %s", connectMaxTries, err)
-			return pnd, err
-		}
-
-		tries++
+func NewTokenQueue() *TokenQueue {
+	return &TokenQueue{
+		tokens: make(chan Token),
 	}
 }
 
@@ -241,117 +119,32 @@ func launchCard(cfg *config.UserConfig, state *State, kbd input.Keyboard) error 
 	return nil
 }
 
-func pollLoop(
-	cfg *config.UserConfig,
-	state *State,
-	kbd input.Keyboard,
-) {
-	var pnd nfc.Device
-	var err error
-
-	ttp := TimesToPoll
-	pbp := PeriodBetweenPolls
-
-	var lastError time.Time
-	playFail := func() {
-		if time.Since(lastError) > 1*time.Second {
-			mister.PlayFail(cfg)
-		}
-	}
-
-	if cfg.TapTo.ExitGame {
-		// FIXME: this method makes the activity indicator flicker, is there another way?
-		ttp = 1
-		// TODO: value requires investigation, originally set to 150 which worked for pn532
-		//       but not for acr122u (read once then never again). 200 seems to work ok
-		pbp = 200 * time.Millisecond
-	}
-
+func processLaunchQueue(cfg *config.UserConfig, state *State, tq *TokenQueue, kbd input.Keyboard) {
 	for {
-		if state.ShouldStopService() {
-			break
-		}
+		select {
+		case t := <-tq.tokens:
+			state.SetActiveCard(t)
 
-		time.Sleep(periodBetweenLoop)
+			if state.IsLauncherDisabled() {
+				continue
+			}
 
-		if connected, _ := state.GetReaderStatus(); !connected {
-			// TODO: keep track of reconnect attempts?
-			log.Info().Msg("reader not connected, attempting connection....")
-
-			pnd, err = OpenDeviceWithRetries(cfg.TapTo, state)
+			err := launchCard(cfg, state, kbd)
 			if err != nil {
-				continue
+				log.Error().Msgf("error launching card: %s", err)
 			}
-
-			if err := pnd.InitiatorInit(); err != nil {
-				state.SetReaderDisconnected()
-				log.Error().Msgf("could not init initiator: %s", err)
-				continue
+		case <-time.After(1 * time.Second):
+			if state.ShouldStopService() {
+				tq.Close()
+				return
 			}
-
-			log.Info().Msgf("opened connection: %s %s", pnd, pnd.Connection())
 		}
-
-		activeCard := state.GetActiveCard()
-
-		log.Debug().Msgf("polling for %d times with %s delay", ttp, pbp)
-		newScanned, removed, err := pollDevice(cfg, &pnd, activeCard, ttp, pbp)
-
-		if errors.Is(err, nfc.Error(nfc.EIO)) {
-			state.SetReaderDisconnected()
-			log.Error().Msgf("error during poll: %s", err)
-			log.Error().Msg("fatal IO error, device was possibly unplugged")
-			playFail()
-			lastError = time.Now()
-			continue
-		} else if err != nil {
-			log.Error().Msgf("error during poll: %s", err)
-			playFail()
-			lastError = time.Now()
-			continue
-		}
-
-		state.SetActiveCard(newScanned)
-
-		if removed && cfg.TapTo.ExitGame && !inExitGameBlocklist(cfg) && !state.IsLauncherDisabled() {
-			mister.ExitGame()
-			continue
-		}
-
-		if newScanned.UID == "" || activeCard.UID == newScanned.UID {
-			continue
-		}
-
-		mister.PlaySuccess(cfg)
-
-		err = writeScanResult(newScanned)
-		if err != nil {
-			log.Error().Msgf("error writing tmp scan result: %s", err)
-		}
-
-		if state.IsLauncherDisabled() {
-			log.Info().Msg("launcher disabled, skipping command action")
-			continue
-		}
-
-		err = launchCard(cfg, state, kbd)
-		if err != nil {
-			log.Error().Msgf("error launching card: %s", err)
-			playFail()
-			lastError = time.Now()
-			continue
-		}
-	}
-
-	state.SetReaderDisconnected()
-	err = pnd.Close()
-	if err != nil {
-		log.Warn().Msgf("error closing device: %s", err)
 	}
 }
 
 func StartDaemon(cfg *config.UserConfig) (func() error, error) {
 	state := &State{}
+	tq := NewTokenQueue()
 
 	// TODO: this is platform specific
 	kbd, err := input.NewKeyboard()
@@ -379,9 +172,9 @@ func StartDaemon(cfg *config.UserConfig) (func() error, error) {
 		state.DisableLauncher()
 	}
 
-	go runApiServer(cfg, state, kbd)
-
-	go pollLoop(cfg, state, kbd)
+	go runApiServer(cfg, state, tq, kbd)
+	go readerPollLoop(cfg, state, tq, kbd)
+	go processLaunchQueue(cfg, state, tq, kbd)
 
 	socket, err := StartSocketServer(state)
 	if err != nil {
@@ -394,6 +187,7 @@ func StartDaemon(cfg *config.UserConfig) (func() error, error) {
 		if err != nil {
 			log.Warn().Msgf("error closing socket: %s", err)
 		}
+		tq.Close()
 		state.StopService()
 		if closeDbWatcher != nil {
 			return closeDbWatcher()
